Quivers

Quivers are represented via their adjacency matrix. Vertices are numbered from $1$ to $n$.

QuiverTools.Quiver — Type

Summary

struct Quiver{T}

A quiver is represented by its adjacency $n \times n$ matrix $adjacency = (a_{ij})$,
where $n$ is the number of vertices and $a_{ij}$ is the number of arrows $i \to j$.

Fields

adjacency :: AbstractMatrix{Int}
name :: String

For all built-in quivers and the different ways to construct them, see Constructors.

Properties

QuiverTools offers various methods to study some graph-theoretical properties of quivers.

QuiverTools.n_vertices — Function

n_vertices(Q::Quiver)

Return the number of vertices of Q.

Examples

julia> Q = kronecker_quiver(4);

julia> n_vertices(Q) == 2
true

QuiverTools.n_arrows — Function

n_arrows(Q::Quiver)

Return the number of arrows of Q.

Examples

julia> Q = kronecker_quiver(4);

julia> n_arrows(Q) == 4
true

QuiverTools.arrows — Function

arrows(Q::Quiver)

Return a list of all arrows of Q.

Examples

julia> arrows(kronecker_quiver(3))
3-element Vector{Tuple{Int64, Int64}}:
 (1, 2)
 (1, 2)
 (1, 2)

julia> arrows(loop_quiver(3))
3-element Vector{Tuple{Int64, Int64}}:
 (1, 1)
 (1, 1)
 (1, 1)

QuiverTools.indegree — Function

indegree(Q::Quiver, j::Int)

Return the number of incoming arrows to the vertex j.

Examples

julia> Q = kronecker_quiver(4);

julia> indegree(Q, 1)
0

julia> indegree(Q, 2)
4

QuiverTools.outdegree — Function

outdegree(Q::Quiver, i::Int)

Return the number of outgoing arrows from the vertex i.

Examples

julia> Q = kronecker_quiver(4);

julia> outdegree(Q, 1)
4

julia> outdegree(Q, 2)
0

QuiverTools.is_acyclic — Function

is_acyclic(Q::Quiver)

Check whether Q is acyclic, i.e., has no oriented cycles.

Examples

julia> Q = kronecker_quiver(4);

julia> is_acyclic(Q)
true

QuiverTools.is_connected — Function

is_connected(Q::Quiver)

Check whether Q is connected.

Examples

julia> is_connected(Quiver([0 1 0; 0 0 1; 1 0 0]))
true

julia> is_connected(Quiver([0 1 0; 1 0 0; 0 0 2]))
false

julia> is_connected(kronecker_quiver(4))
true

julia> is_connected(loop_quiver(4))
true

julia> is_connected(subspace_quiver(4))
true

QuiverTools.is_sink — Function

is_sink(Q::Quiver, j::Int)

Checks if the vertex j is a sink, i.e., a vertex with no outgoing arrows.

Examples

julia> Q = kronecker_quiver(4);

julia> is_sink(Q, 1)
false

julia> is_sink(Q, 2)
true

QuiverTools.is_source — Function

is_source(Q::Quiver, i::Int)

Check whether the vertex i is a source, i.e., a vertex with no incoming arrows.

Examples

julia> Q = kronecker_quiver(4);

julia> is_source(Q, 1)
true

julia> is_source(Q, 2)
false

QuiverTools.underlying_graph — Function

underlying_graph(Q::Quiver)

Return the (necessarily symmetric) adjacency matrix of the underlying graph of Q.

Examples

julia> Q = kronecker_quiver(4);

julia> underlying_graph(Q) == [0 4; 4 0]
true