Methods

Here are listed all the available methods of QuiverTools

QuiverTools.Quiver — Type

A quiver is represented by its adjacency $n \times n$ matrix $adjacency = (a_{ij})$, where $n$ is the number of vertices and $a_{ij}$ is the number of arrows $i \to j$.

Attributes:

adjacency is the adjacency matrix of the quiver

name is the name of the quiver, defaults to "".

QuiverTools.CardinalGd — Method

Cardinality of product of general linear groups $\mathrm{GL}_{d}(\mathbb{F}_q)$.

QuiverTools.CardinalGl — Method

Cardinality of general linear group $\mathrm{GL}_n(\mathbb{F}_v)$.

QuiverTools.CardinalRd — Method

Cardinality of representation space $\mathrm{R}(Q,d)$, over $\mathbb{F}_q$.

QuiverTools.Euler_form — Method

Computes the Euler form of the quiver for vectors $x$ and $y$.

The Euler form is defined as the bilinear form

\[\langle x,y\rangle = x^T * E * y,\]

where $E$ is the Euler matrix of the quiver.

QuiverTools.Euler_matrix — Method

Returns the Euler matrix of the quiver.

The Euler matrix of a quiver $Q$ is defined as

\[E = I - A,\]

where $A$ is the adjacency matrix of $Q$ and $I$ is the identity matrix of the same size as $A$.

QuiverTools.Hodge_diamond — Method

Returns the Hodge diamond of the moduli space of $\theta$-semistable representations of $Q$ with dimension vector $d$.

QuiverTools.Hodge_polynomial — Method

Returns the Hodge polynomial of the moduli space of $\theta$-semistable representations of $Q$ with dimension vector $d$.

The algorithm is based on MR1974891, and the current implementation is translated from the Hodge diamond cutter.

QuiverTools.Picard_rank — Method

Computes the Picard rank of the moduli space of $\theta$-semistable representations of $Q$ with dimension vector $d$.

QuiverTools.Teleman_bound_onstratum — Function

Computes the weight on $\det(N_{S/R}|_Z)$ of the 1-PS $\lambda$ corresponding to the given HN type.

QuiverTools.TransferMatrixEntry — Method

Entry of the transfer matrix, as per Corollary 6.9

QuiverTools.all_HN_types — Function

Returns a list of all the Harder Narasimhan types of representations of $Q$ with dimension vector $d$, with respect to the slope function theta/slope_denominator.

Examples:

julia> Q = mKronecker_quiver(3); d = [2,3]; theta = [3,-2];

julia> all_HN_types(Q, d, theta)
8-element Vector{Vector{Vector{Int64}}}:
 [[2, 3]]
 [[1, 1], [1, 2]]
 [[2, 2], [0, 1]]
 [[2, 1], [0, 2]]
 [[1, 0], [1, 3]]
 [[1, 0], [1, 2], [0, 1]]
 [[1, 0], [1, 1], [0, 2]]
 [[2, 0], [0, 3]]

julia> all_HN_types(Q, [3,0], [0,0])
1-element Vector{Vector{Vector{Int64}}}:
 [[3, 0]]

julia> Q = three_vertex_quiver(1,4,1); d = [4,1,4];

julia> theta = canonical_stability(Q, d);

julia> all_HN_types(Q, d, theta)
106-element Vector{Vector{Vector{Int64}}}:
 [[4, 1, 4]]
 [[4, 1, 3], [0, 0, 1]]
 [[4, 0, 3], [0, 1, 1]]
 [[4, 0, 3], [0, 1, 0], [0, 0, 1]]
 [[3, 1, 2], [1, 0, 2]]
 [[3, 1, 2], [1, 0, 1], [0, 0, 1]]
 [[3, 0, 2], [1, 1, 2]]
 [[3, 0, 2], [0, 1, 0], [1, 0, 2]]
 [[3, 0, 2], [1, 0, 1], [0, 1, 1]]
 [[3, 0, 2], [1, 1, 1], [0, 0, 1]]
 ⋮
 [[3, 0, 0], [1, 1, 2], [0, 0, 2]]
 [[3, 0, 0], [0, 1, 0], [1, 0, 4]]
 [[3, 0, 0], [0, 1, 0], [1, 0, 3], [0, 0, 1]]
 [[3, 0, 0], [0, 1, 0], [1, 0, 2], [0, 0, 2]]
 [[3, 0, 0], [1, 0, 1], [0, 1, 1], [0, 0, 2]]
 [[3, 0, 0], [1, 1, 1], [0, 0, 3]]
 [[3, 0, 0], [1, 1, 0], [0, 0, 4]]
 [[4, 0, 0], [0, 1, 1], [0, 0, 3]]
 [[4, 0, 0], [0, 1, 0], [0, 0, 4]]

QuiverTools.all_Teleman_bounds — Function

Computes the weight on $\det(N_{S/R}|_Z)$ of the 1-PS corresponding to each HN type for the given $Q$, $d$, $\theta$ and slope_denominator`.

QuiverTools.all_destabilizing_subdimension_vectors — Function

Returns the subdimension vectors of $d$ with a strictly larger slope than $d$.

QuiverTools.all_generic_subdimension_vectors — Method

Returns the list of all generic subdimension vectors of $d$.

Examples:

julia> Q = mKronecker_quiver(3); d = [2,3];

julia> QuiverTools.all_generic_subdimension_vectors(Q, d)
7-element Vector{Vector{Int64}}:
 [0, 0]
 [0, 1]
 [0, 2]
 [1, 2]
 [0, 3]
 [1, 3]
 [2, 3]

QuiverTools.all_slope_decreasing_sequences — Function

Returns the list of all sequences $(d^1,...,d^l)$ which sum to $d$ such that $\mu(d^1) > ... > \mu(d^l).$

Examples:

julia> Q = mKronecker_quiver(3); d = [2,3]; theta = [3,-2];

julia> QuiverTools.all_slope_decreasing_sequences(Q, d, theta)
8-element Vector{Vector{Vector{Int64}}}:
 [[2, 3]]
 [[1, 1], [1, 2]]
 [[2, 2], [0, 1]]
 [[2, 1], [0, 2]]
 [[1, 0], [1, 3]]
 [[1, 0], [1, 2], [0, 1]]
 [[1, 0], [1, 1], [0, 2]]
 [[2, 0], [0, 3]]

QuiverTools.all_weights_endomorphisms_universal_bundle — Function

Computes the weights of the endomorphisms of the universal bundles $U_i \otimes U_j$ on all the non-dense Harder-Narasimhan strata for each 1-PS relative to the HN type.

QuiverTools.all_weights_irreducible_component_canonical — Function

Computes the weights of the irreducible component of $\omega_R|_Z$ on all the non-dense Harder-Narasimhan strata for each 1-PS relative to the HN type. More explicitly, if $\omega_X = O(rH)$, this returns the weights of the pullback of $\mathcal{O}(H)$ on each stratum.

QuiverTools.all_weights_universal_bundle — Function

Computes the weights of the universal bundle $U_i(a)$ for the linearization $a$ on all the non-dense Harder-Narasimhan strata for each 1-PS corresponding to each HN type.

QuiverTools.canonical_decomposition — Method

Computes the canonical decomposition of the dimension vector $d$ for the given quiver $Q$.

If $\beta_1, \dots, \beta_{\ell}$ is a sequence of Schur roots such that, for all $i \neq j$, one has

\[\mathrm{ext}(\beta_i, \beta_j) = \mathrm{ext}(\beta_j, \beta_i) = 0,\]

then the general representation of dimension $\sum_i \beta_i$ is isomorphic to the direct sum of irreducible representations of dimension vectors $\beta_i$.

Such a decomposition is called the canonical decomposition.

QuiverTools.canonical_stability — Method

The canonical stability parameter for the couple $(Q,d)$ is given by $<d,- > - < - ,d>$

QuiverTools.codimension_HN_stratum — Method

Returns the codimension of the given HN stratum.

Examples:

julia> Q = mKronecker_quiver(3); d = [2,3]; theta = [3,-2];

julia> HN = all_HN_types(Q, d, theta)
8-element Vector{Vector{Vector{Int64}}}:
 [[2, 3]]
 [[1, 1], [1, 2]]
 [[2, 2], [0, 1]]
 [[2, 1], [0, 2]]
 [[1, 0], [1, 3]]
 [[1, 0], [1, 2], [0, 1]]
 [[1, 0], [1, 1], [0, 2]]
 [[2, 0], [0, 3]]

julia> for hntype in HN
        println(QuiverTools.codimension_HN_stratum(Q, hntype))
        end
0
3
4
10
8
9
12
18

QuiverTools.double_quiver — Method

The adjacency matrix of the double of a quiver is the sum of the adjacency matrix of the original quiver and its transpose.

QuiverTools.extension_matrix — Method

Given an HN type $(d^1,...,d^l)$ for the quiver Q, returns the upper triangular matrix whose $i,j$-th entry is $\mathrm{ext}(d^i,d^j)$.

The sum of all the entries is the codimension of the HN stratum; the sum of all the rectangles starting on the "up-diagonal" (where the 1s go in a Jordan form) and going all the way to the entry $1,\ell$ is at least $1$.

The Teleman inequality is satisfied for this stratum iif one of these rectangles sums to $2$ or more.

QuiverTools.generic_ext — Method

Computes the dimension of the $\mathrm{Ext}^1$ group between generic representations of dimension vectors $a$ and $b$.

According to Thm. 5.4 in Schofield's General representations of quivers, we have

\[ext(a,b) = max\{- \langle c, b\rangle~~|~~c~\text{is a generic subdimension vector of } a\}.\]

QuiverTools.generic_hom — Method

Computes the dimension of the $\mathrm{Hom}$ group between generic representations of dimension vectors $a$ and $b$.

QuiverTools.has_semistables — Function

Checks if there is a $\theta$-semistable representation of dimension vector $d$.

Examples:

julia> A2 = mKronecker_quiver(1); theta = [1,-1];

julia> has_semistables(A2, [1,1], theta)
true

julia> has_semistables(A2, [2,2], theta)
true

julia> has_semistables(A2, [1,2], theta)
false

julia> has_semistables(A2, [0,0], theta)
true

julia> # The 3-Kronecker quiver:

julia> K3 = mKronecker_quiver(3); theta = [3,-2];

julia> has_semistables(K3, [2,3], theta)
true

julia> has_semistables(K3, [1,4], theta)
false

QuiverTools.has_stables — Function

Checks if Q has a $theta$-stable representation of dimension vector $d$.

Examples:

julia> Q = mKronecker_quiver(3); d = [2,3]; theta = [3,-2];

julia> has_stables(Q, d, theta)
true

julia> Q = mKronecker_quiver(2); d = [2,2]; theta = [1,-1];

julia> has_stables(Q, d, theta)
false

julia> has_semistables(Q, d, theta)
true

QuiverTools.in_fundamental_domain — Method

Checks if the dimension vector $d$ is in the fundamental domain of the quiver $Q$.

The fundamental domain is the cone of dimension vectors in $\mathbb{Z}^{Q_0}$ such that the symmetric Tits form is negative on all the simple roots, i.e., for all vertices i,

\[(s_i, d) := \langle d, s_i\rangle + \langle s_i, d\rangle \leq 0,\]

where $s_i$ is the dimension vector with all entries set to $0$ and the i-th set to $1$.

QuiverTools.in_stable_cone — Function

Checks if the stability parameter $\theta$ belongs to the cone of parameters admitting stable representations of dimension vector $d$. Assumes that the dimension vector $d$ is Schurian (for now).

QuiverTools.indegree — Method

Returns the number of incoming arrows to the vertex $j$.

Examples:

julia> Q = mKronecker_quiver(4);

julia> indegree(Q, 1)
0

julia> indegree(Q, 2)
4

QuiverTools.is_Schur_root — Method

Checks if $d$ is a Schur root for $Q$.

By a lemma of Schofield (See Lemma 4.2 of arXiv:0802.2147), this is equivalent to the existence of a stable representation of dimension vector $d$ for the canonical stability parameter.

Examples:

julia> Q = mKronecker_quiver(3); d = [2,3];

julia> is_Schur_root(Q, d)
true

QuiverTools.is_acyclic — Method

Checks wether the quiver is acyclic, i.e. has no oriented cycles.

QuiverTools.is_amply_stable — Function

Checks wether the dimension vector $d$ is amply stable with respect to the slope function theta/denominator.

This means that the codimension of the unstable locus in the parameter space is at least $2$.

QuiverTools.is_connected — Method

Checks wether the underlying graph of the quiver is connected.

Examples:

julia> Q = Quiver([0 1 0; 0 0 1; 1 0 0]);

julia> is_connected(Q)
true

julia> Q = Quiver([0 1 0; 1 0 0; 0 0 2]);

julia> is_connected(Q)
false

julia> # The 4-Kronecker quiver:

julia> Q = mKronecker_quiver(4);

julia> is_connected(Q)
true

julia> # The 4-loop quiver:

julia> Q = loop_quiver(4);

julia> is_connected(Q)
true

julia> # The 4-subspace quiver:

julia> Q = subspace_quiver(4);

julia> is_connected(Q)
true

julia> # The A10 quiver:

julia> A10 = Quiver(   [0 1 0 0 0 0 0 0 0 0;
                        0 0 1 0 0 0 0 0 0 0;
                        0 0 0 1 0 0 0 0 0 0;
                        0 0 0 0 1 0 0 0 0 0;
                        0 0 0 0 0 1 0 0 0 0;
                        0 0 0 0 0 0 1 0 0 0;
                        0 0 0 0 0 0 0 1 0 0;
                        0 0 0 0 0 0 0 0 1 0;
                        0 0 0 0 0 0 0 0 0 1;
                        0 0 0 0 0 0 0 0 0 0] );

julia> is_connected(A10)
true

julia> # The A10 quiver without one arrow:

julia> A10 = Quiver(   [0 1 0 0 0 0 0 0 0 0;
                        0 0 1 0 0 0 0 0 0 0;
                        0 0 0 1 0 0 0 0 0 0;
                        0 0 0 0 1 0 0 0 0 0;
                        0 0 0 0 0 1 0 0 0 0;
                        0 0 0 0 0 0 0 0 0 0;
                        0 0 0 0 0 0 0 1 0 0;
                        0 0 0 0 0 0 0 0 1 0;
                        0 0 0 0 0 0 0 0 0 1;
                        0 0 0 0 0 0 0 0 0 0] );

julia> is_connected(A10)
false

QuiverTools.is_coprime — Method

Checks wether the given dimension vector $d$ is $\theta$-coprime for the stability parameter $\theta$.

QuiverTools.is_coprime — Method

Checks if the gcd of all the entries of d is $1$.

QuiverTools.is_generic_subdimension_vector — Method

Checks if $e$ is a generic subdimension vector of $d$.

A dimension vector $e$ is called a generic subdimension vector of $d$ if a generic representation of dimension vector $d$ possesses a subrepresentation of dimension vector $e$.

By a result of Schofield (see Thm. 5.3 of arXiv:0802.2147), $e$ is a generic subdimension vector of $d$ if and only if

\[<e',d-e> \geq 0\]

for all generic subdimension vectors $e'$ of $e$.

QuiverTools.is_sink — Method

Checks if the vertex $j$ is a sink, i.e., a vertex with no outgoing arrows.

Examples:

julia> Q = mKronecker_quiver(4);

julia> is_sink(Q, 1)
false

julia> is_sink(Q, 2)
true

QuiverTools.is_source — Method

Checks if the vertex $i$ is a source, i.e., a vertex with no incoming arrows.

Examples:

julia> Q = mKronecker_quiver(4);

julia> is_source(Q, 1)
true

julia> is_source(Q, 2)
false

QuiverTools.narrows — Method

Returns the number of arrows of the quiver.

QuiverTools.nvertices — Method

Returns the number of vertices of the quiver.

QuiverTools.opposite_quiver — Method

" Returns a Quiver with the same vertices and an arrow $j \to i$ for every arrow $i \to j$ in the original quiver.

QuiverTools.outdegree — Method

Returns the number of outgoing arrows from the vertex $i$.

Examples:

julia> Q = mKronecker_quiver(4);

julia> outdegree(Q, 1)
4

julia> outdegree(Q, 2)
0

QuiverTools.slope — Function

Returns the slope of the dimension vector $d$ with respect to the stability parameter $\theta$ and a choice of a denominator function.

QuiverTools.smooth_model_quiver — Method

Computes the quiver $\hat{Q}$ defined in arXiv:0706.4306 for the given quiver $Q$ and a given subdimension vector $e$.

QuiverTools.solve — Method

Solve $A\cdot x = b$ for $A$ upper triangular via back substitution

QuiverTools.symmetric_polynomial — Method

Returns the symmetric polynomial of degree $degree$ in the variables $vars$.

QuiverTools.underlying_graph — Method

Returns the (necessarily symmetric) adjacency matrix of the underlying graph of the quiver.

QuiverTools.weight_irreducible_component_canonical_on_stratum — Function

Computes the weight of the irreducible component of $\omega_R|_Z$ on a Harder-Narasimhan stratum for the 1-PS corresponding to each HN type. More explicitly, if $\omega_X = \mathcal{O}(rH)$, this returns the weight of the pullback of O(H) on the given stratum.

QuiverTools.weights_endomorphism_universal_bundle_on_stratum — Function

Computes the weights of the endomorphism of the universal bundle $U_i \otimes U_j$ on the given Harder-Narasimhan stratum for the 1-PS relative to the HN type.

QuiverTools.weights_universal_bundle_onstratum — Function

Returns the weights of a universal bundle $U_i(a)$ for the linearization $a$ for the 1-PS corresponding to the given HN type.