Methods

Here are listed all the available methods of QuiverTools

QuiverTools.Quiver — Type

A quiver is represented by its adjacency $n \times n$ matrix $adjacency = (a_{ij})$, where $n$ is the number of vertices and $a_{ij}$ is the number of arrows $i \to j$.

Attributes:

adjacency is the adjacency matrix of the quiver

name is the name of the quiver, defaults to "".

QuiverTools.Betti_numbers — Method

Betti_numbers(M::QuiverModuliSpace)

Returns the Betti numbers of the moduli space $M$.

INPUT:

$M$: a moduli space of representations of a quiver.

OUTPUT:

a list of Betti numbers of the moduli space.

EXAMPLES:

julia> Q = mKronecker_quiver(2);

julia> M = QuiverModuliSpace(Q, [1, 1]);

julia> Betti_numbers(M)
3-element Vector{Int64}:
 1
 0
 1

Our favourite 6-fold:

julia> Q = mKronecker_quiver(3);

julia> M = QuiverModuliSpace(Q, [2, 3]);

julia> Betti_numbers(M)
13-element Vector{Int64}:
 1
 0
 1
 0
 3
 0
 3
 0
 3
 0
 1
 0
 1

QuiverTools.CardinalGd — Method

Cardinality of product of general linear groups $\mathrm{GL}_{d}(\mathbb{F}_q)$.

QuiverTools.CardinalGl — Method

Cardinality of general linear group $\mathrm{GL}_n(\mathbb{F}_v)$.

QuiverTools.CardinalRd — Method

Cardinality of representation space $\mathrm{R}(Q,d), over \mathbb{F}_q$.

QuiverTools.Chern_character_line_bundle — Method

Chern_character_line_bundle(M::QuiverModuliSpace, eta)

Returns the Chern character of the line bundle L(eta).

INPUT:

$M$: a moduli space of representations of a quiver.

$eta$: a choice of linearization for the trivial line bundle.

OUTPUT:

the Chern character of the line bundle L(eta).

EXAMPLES:

Some line bundles on the projective line:

julia> Q = mKronecker_quiver(2); M = QuiverModuliSpace(Q, [1, 1]);

julia> Chern_character_line_bundle(M, [1, -1])
-x11 + 1

Some Chern characters for our favourite 6-fold:

julia> Q = mKronecker_quiver(3); M = QuiverModuliSpace(Q, [2, 3]);

julia> Chern_character_line_bundle(M, [3, -2])
1//720*x21^6 - 1//120*x21^5 + 1//24*x21^4 - 1//6*x21^3 + 1//2*x21^2 - x21 + 1

QuiverTools.Chern_class_line_bundle — Method

Chern_class_line_bundle(M::QuiverModuliSpace, eta)

Returns the first Chern class of the line bundle L(eta).

This is given by $L(eta) = \bigoplus_{i \in Q_0} \det(U_i)^{-eta_i}$.

INPUT:

$M$: a moduli space of representations of a quiver.

$eta$: a choice of linearization for the trivial line bundle.

OUTPUT:

the first Chern class of the line bundle L(eta) as a polynomial.

EXAMPLES:

The line bundles $\mathcal{O}(i)$ on the projective line:

julia> Q = mKronecker_quiver(2); M = QuiverModuliSpace(Q, [1, 1]);

julia> l = Chern_class_line_bundle(M, [1, -1])
-x11

The line bundle corresponding to the canonical stability condition on our favourite 6-fold:

julia> Q = mKronecker_quiver(3); M = QuiverModuliSpace(Q, [2, 3]);

julia> Chern_class_line_bundle(M, [9, -6])
-3*x21

QuiverTools.Chow_ring — Function

Chow_ring(Q, d, theta, a)

Computes the Chow ring of the moduli space of $\theta$-semistable representations of $Q$ with dimension vector $d$, for a choice of linearization $a$.

This method of the function Chow_ring also returns the ambient ring $R$ and the inclusion morphism.

INPUT:

$Q$: a quiver.

$d$: a dimension vector.

$theta$: a stability parameter. Default is the canonical stability.

$a$: a linearization. Default is the extended gcd of $d$.

OUTPUT: A tuple containing:

the Chow ring of the moduli space,

the polynomial ring above it,

the inclusion map $\iota : A \to R$.

EXAMPLES:

The Chow ring for the projective line has two generators:

julia> Q = mKronecker_quiver(2); M = QuiverModuliSpace(Q, [1, 1]);

julia> CH = Chow_ring(M)[1];

julia> QuiverTools.gens(QuiverTools.quotient_ideal(CH))
2-element Vector{Singular.spoly{Singular.n_Q}}:
 x21
 x11^2

The Chow ring for our favourite 6-fold has, in this implementation, 16 generators:

julia> Q = mKronecker_quiver(3); M = QuiverModuliSpace(Q, [2, 3]);

julia> CH = Chow_ring(M)[1]; I = QuiverTools.quotient_ideal(CH);

julia> length(QuiverTools.gens(I))
16

QuiverTools.Chow_ring — Function

Chow_ring(M::QuiverModuliSpace, chi)

Computes the Chow ring of the moduli space $M$.

INPUT:

$M$: a moduli space of representations of a quiver.

$chi$: a choice of linearization for the trivial line bundle. It picks one by default if not provided.

OUTPUT:

the Chow ring of the moduli space.

QuiverTools.Dynkin_quiver — Method

Dynkin_quiver(T, n)

Constructs the Dynkin quiver, with arbitrary orientation of the arrows.

EXAMPLES:

julia> Dynkin_quiver("D4")
Dynkin quiver of type D4, with adjacency matrix [0 1 0 0; 0 0 1 1; 0 0 0 0; 0 0 0 0]

QuiverTools.Euler_form — Method

Computes the Euler form of the quiver for vectors $x$ and $y$.

The Euler form is defined as the bilinear form

\[\langle x,y\rangle = x^T * E * y,\]

where $E$ is the Euler matrix of the quiver.

EXAMPLE:

julia> Q = mKronecker_quiver(4);

julia> Euler_form(Q, [1, 1], [1, 1]) == -2
true

QuiverTools.Euler_matrix — Method

Returns the Euler matrix of the quiver.

The Euler matrix of a quiver $Q$ is defined as

\[E = I - A,\]

where $A$ is the adjacency matrix of $Q$ and $I$ is the identity matrix of the same size as $A$.

EXAMPLE:

julia> Q = mKronecker_quiver(4);

julia> QuiverTools.Euler_matrix(Q) == [1 -4; 0 1]
true

QuiverTools.Hodge_diamond — Function

Hodge_diamond(Q, d, theta)

Returns the Hodge diamond of the moduli space of $\theta$-semistable representations of $Q$ with dimension vector $d$.

INPUT:

$Q$: a quiver.

$d$: a dimension vector.

$theta$: a stability parameter. Default is the canonical stability.

OUTPUT:

the Hodge diamond of the moduli space.

EXAMPLES:

The Hodge diamond of our favourite 6-fold:

julia> Q = mKronecker_quiver(3);

julia> Hodge_diamond(Q, [2, 3])
7×7 Matrix{Int64}:
 1  0  0  0  0  0  0
 0  1  0  0  0  0  0
 0  0  3  0  0  0  0
 0  0  0  3  0  0  0
 0  0  0  0  3  0  0
 0  0  0  0  0  1  0
 0  0  0  0  0  0  1

QuiverTools.Hodge_diamond — Method

Hodge_diamond(M::QuiverModuliSpace)

Returns the Hodge diamond of the moduli space $M$.

INPUT:

$M$: a moduli space of representations of a quiver.

OUTPUT:

the Hodge diamond of the moduli space.

EXAMPLES:

The Hodge diamond of our favourite 6-fold:

julia> Q = mKronecker_quiver(3);

julia> M = QuiverModuliSpace(Q, [2, 3]);

julia> Hodge_diamond(M)
7×7 Matrix{Int64}:
 1  0  0  0  0  0  0
 0  1  0  0  0  0  0
 0  0  3  0  0  0  0
 0  0  0  3  0  0  0
 0  0  0  0  3  0  0
 0  0  0  0  0  1  0
 0  0  0  0  0  0  1

QuiverTools.Hodge_polynomial — Function

Hodge_polynomial(Q, d, theta)

Returns the Hodge polynomial of the moduli space of $\theta$-semistable representations of $Q$ with dimension vector $d$.

The algorithm is based on MR1974891, and the current implementation is translated from the Hodge diamond cutter.

INPUT:

$Q$: a quiver.

$d$: a dimension vector.

$theta$: a stability parameter. Default is the canonical stability.

OUTPUT:

the Hodge polynomial of the moduli space.

EXAMPLES:

The Hodge polynomial of our favourite 6-fold:

julia> Q = mKronecker_quiver(3);

julia> d = [2, 3];

julia> theta = [3, -2];

julia> Hodge_polynomial(Q, d, theta)
x^6*y^6 + x^5*y^5 + 3*x^4*y^4 + 3*x^3*y^3 + 3*x^2*y^2 + x*y + 1

QuiverTools.Hodge_polynomial — Method

Hodge_polynomial(M::QuiverModuliSpace)

Returns the Hodge polynomial of the moduli space $M$.

INPUT:

$M$: a moduli space of representations of a quiver.

OUTPUT:

the Hodge polynomial of the moduli space.

EXAMPLES:

The Hodge polynomial of our favourite 6-fold:

julia> Q = mKronecker_quiver(3);

julia> M = QuiverModuliSpace(Q, [2, 3]);

julia> Hodge_polynomial(M)
x^6*y^6 + x^5*y^5 + 3*x^4*y^4 + 3*x^3*y^3 + 3*x^2*y^2 + x*y + 1

QuiverTools.Picard_rank — Method

Picard_rank(M::QuiverModuliSpace)

Returns the Picard rank of the moduli space $M$.

INPUT:

$M$: a moduli space of representations of a quiver.

OUTPUT:

the Picard rank of the moduli space.

EXAMPLES:

Kronecker quiver with dimension vector [2, 3]:

julia> Q = mKronecker_quiver(3);

julia> M = QuiverModuliSpace(Q, [2, 3]);

julia> Picard_rank(M)
1

QuiverTools.Poincare_polynomial — Method

Poincare_polynomial(M::QuiverModuliSpace)

Returns the Poincaré polynomial of the moduli space $M$.

INPUT:

$M$: a moduli space of representations of a quiver.

OUTPUT:

the Poincaré polynomial of the moduli space.

EXAMPLES:

A Kronecker quiver setup where $M$ is the projective line:

julia> Q = mKronecker_quiver(2);

julia> M = QuiverModuliSpace(Q, [1, 1]);

julia> Poincare_polynomial(M)
L + 1

The Poincaré polynomial of our favourite 6-fold:

julia> Q = mKronecker_quiver(3);

julia> M = QuiverModuliSpace(Q, [2, 3]);

julia> Poincare_polynomial(M)
L^6 + L^5 + 3*L^4 + 3*L^3 + 3*L^2 + L + 1

QuiverTools.Teleman_bound_onstratum — Function

Computes the weight on $\det(N_{S/R}|_Z)$ of the 1-PS $\lambda$ corresponding to the given HN type.

QuiverTools.Todd_class — Function

todd_class(M::QuiverModuliSpace, chi)

Returns the Todd class of the moduli space $M$.

INPUT:

$M$: a moduli space of representations of a quiver.

$chi$: a choice of linearization to construct the universal bundles.

OUTPUT:

the Todd class of the moduli space, as a polynomial in its Chow ring.

EXAMPLES: The Todd class of our favourite 3-Kronecker quiver moduli:

julia> Q = mKronecker_quiver(3); M = QuiverModuliSpace(Q, [2, 3]);

julia> Todd_class(M)
-17//8*x12*x21 + x21^2 + 823//360*x12*x22 - 823//1080*x22^2 + 553//1080*x21*x23 - 77//60*x22*x23 + x23^2 + 5//12*x12 - 3//2*x21 + 9//8*x23 + 1

QuiverTools.TransferMatrixEntry — Method

Entry of the transfer matrix, as per Corollary 6.9 of MR1974891

QuiverTools.__Chow_degrees — Method

__Chow_degrees(d)

Returns the vector of degrees for the variables of a Chow ring.

For internal use only.

QuiverTools.__Chow_ring_monomial_grading — Method

__Chow_ring__monomial_grading(M, f)

Returns the "pseudodegree" of the monomial f in the Chow ring of the moduli space M passed.

This method is unsafe, as it does not consider the actual degree of the MPolyRingElem objects passed. Instead, it assumes that the Chow ring passed has variables $x_{i, j}$ as in the Chow ring paper.

QuiverTools.all_HN_types — Function

all_HN_types(Q, d, theta, denom, ordered=true)

Returns a list of all the Harder Narasimhan types of representations of $Q$ with dimension vector $d$, with respect to the slope function theta/denom.

EXAMPLES:

julia> Q = mKronecker_quiver(3); d = [2,3]; theta = [3,-2];

julia> all_HN_types(Q, d, theta)
8-element Vector{Vector{AbstractVector{Int64}}}:
 [[2, 3]]
 [[1, 1], [1, 2]]
 [[2, 2], [0, 1]]
 [[2, 1], [0, 2]]
 [[1, 0], [1, 3]]
 [[1, 0], [1, 2], [0, 1]]
 [[1, 0], [1, 1], [0, 2]]
 [[2, 0], [0, 3]]

julia> all_HN_types(Q, [3,0], [0,0])
1-element Vector{Vector{AbstractVector{Int64}}}:
 [[3, 0]]

julia> Q = three_vertex_quiver(1, 4, 1); d = [4, 1, 4];

julia> theta = canonical_stability(Q, d);

julia> all_HN_types(Q, d, theta)
106-element Vector{Vector{AbstractVector{Int64}}}:
 [[4, 1, 4]]
 [[4, 1, 3], [0, 0, 1]]
 [[4, 0, 3], [0, 1, 1]]
 [[4, 0, 3], [0, 1, 0], [0, 0, 1]]
 [[3, 1, 2], [1, 0, 2]]
 [[3, 1, 2], [1, 0, 1], [0, 0, 1]]
 [[3, 0, 2], [1, 1, 2]]
 [[3, 0, 2], [0, 1, 0], [1, 0, 2]]
 [[3, 0, 2], [1, 0, 1], [0, 1, 1]]
 [[3, 0, 2], [1, 1, 1], [0, 0, 1]]
 ⋮
 [[3, 0, 0], [1, 1, 2], [0, 0, 2]]
 [[3, 0, 0], [0, 1, 0], [1, 0, 4]]
 [[3, 0, 0], [0, 1, 0], [1, 0, 3], [0, 0, 1]]
 [[3, 0, 0], [0, 1, 0], [1, 0, 2], [0, 0, 2]]
 [[3, 0, 0], [1, 0, 1], [0, 1, 1], [0, 0, 2]]
 [[3, 0, 0], [1, 1, 1], [0, 0, 3]]
 [[3, 0, 0], [1, 1, 0], [0, 0, 4]]
 [[4, 0, 0], [0, 1, 1], [0, 0, 3]]
 [[4, 0, 0], [0, 1, 0], [0, 0, 4]]

QuiverTools.all_HN_types — Method

all_HN_types(M::QuiverModuli; unstable::Bool = false, ordered::Bool = true)

Returns all Harder-Narasimhan types of the moduli space.

INPUT:

M::QuiverModuli: a moduli space or stack of representations of a quiver.

unstable::Bool = false: if true, returns only Harder-Narasimhan types

corresponding to unstable representations.

ordered::Bool = true: if true, returns the Harder-Narasimhan types in

the order introduced by MR1974891.

OUTPUT:

a list of Harder-Narasimhan types for the dimension vector and slope of M.

EXAMPLES:

The HN types for a 3-Kronecker quiver with dimension vector [2, 3]:

julia> Q = mKronecker_quiver(3); M = QuiverModuliSpace(Q, [2, 3]);

julia> all_HN_types(M)
8-element Vector{Vector{AbstractVector{Int64}}}:
 [[2, 3]]
 [[1, 1], [1, 2]]
 [[2, 2], [0, 1]]
 [[2, 1], [0, 2]]
 [[1, 0], [1, 3]]
 [[1, 0], [1, 2], [0, 1]]
 [[1, 0], [1, 1], [0, 2]]
 [[2, 0], [0, 3]]

julia> all_HN_types(M, unstable = true)
7-element Vector{Vector{AbstractVector{Int64}}}:
 [[1, 1], [1, 2]]
 [[2, 2], [0, 1]]
 [[2, 1], [0, 2]]
 [[1, 0], [1, 3]]
 [[1, 0], [1, 2], [0, 1]]
 [[1, 0], [1, 1], [0, 2]]
 [[2, 0], [0, 3]]

QuiverTools.all_Luna_types — Function

all_Luna_types(Q, d, theta, denom, exclude_stable)

Computes all the possible Luna types for the given data.

INPUT:

Q: a quiver.

d: a dimension vector.

theta: a stability parameter. Defaults to the canonical stability.

denom: a function defining the denominator of the slope. Defaults to sum.

exclude_stable: whether to exclude the Luna type of stable representations.

OUTPUT: A list of Luna types.

EXAMPLES:

julia> Q = mKronecker_quiver(3); M = QuiverModuliSpace(Q, [3, 3]);

julia> all_Luna_types(M)
5-element Vector{Dict{AbstractVector, Vector{Int64}}}:
 Dict([3, 3] => [1])
 Dict([1, 1] => [1], [2, 2] => [1])
 Dict([1, 1] => [3])
 Dict([1, 1] => [2, 1])
 Dict([1, 1] => [1, 1, 1])

julia> X = QuiverModuliSpace(Q, [2, 3]);

julia> all_Luna_types(X)
1-element Vector{Dict{AbstractVector, Vector{Int64}}}:
 Dict([2, 3] => [1])

QuiverTools.all_Luna_types — Method

all_Luna_types(M::QuiverModuli; exclude_stable::Bool = false)

Returns all Luna types of the moduli space.

INPUT:

M::QuiverModuli: a moduli space or stack of representations of a quiver.

exclude_stable::Bool = false: if true, excludes the stable Luna type.

OUTPUT:

a list of Luna types for the dimension vector and slope of M.

EXAMPLES:

Luna types for a 3-Kronecker quiver:

julia> Q = mKronecker_quiver(3); M = QuiverModuliSpace(Q, [3, 3]);

julia> all_Luna_types(M)
5-element Vector{Dict{AbstractVector, Vector{Int64}}}:
 Dict([3, 3] => [1])
 Dict([1, 1] => [1], [2, 2] => [1])
 Dict([1, 1] => [3])
 Dict([1, 1] => [2, 1])
 Dict([1, 1] => [1, 1, 1])

QuiverTools.all_Teleman_bounds — Function

Computes the weight on $\det(N_{S/R}|_Z)$ of the 1-PS corresponding to each HN type for the given $Q$, $d$, $\theta$ and denom`.

EXAMPLES:

julia> Q = mKronecker_quiver(3);

julia> all_Teleman_bounds(Q, [2, 3], [3, -2])
Dict{Vector{AbstractVector{Int64}}, Int64} with 7 entries:
  [[2, 2], [0, 1]]         => 20
  [[2, 1], [0, 2]]         => 100
  [[1, 0], [1, 2], [0, 1]] => 100
  [[1, 0], [1, 3]]         => 120
  [[1, 0], [1, 1], [0, 2]] => 90
  [[1, 1], [1, 2]]         => 15
  [[2, 0], [0, 3]]         => 90

QuiverTools.all_Teleman_bounds — Method

Interface for all_Teleman_bounds(Q, d, theta).

EXAMPLE:

julia> Q = three_vertex_quiver(1, 2, 3); d = [3, 1, 2]; theta = [5, 3, -9];

julia> all_Teleman_bounds(Q, d, theta)
Dict{Vector{AbstractVector{Int64}}, Int64} with 24 entries:
  [[2, 1, 1], [1, 0, 1]]                       => 12
  [[1, 0, 0], [0, 1, 0], [2, 0, 1], [0, 0, 1]] => 306
  [[1, 0, 0], [1, 1, 0], [1, 0, 1], [0, 0, 1]] => 131
  [[2, 0, 0], [1, 0, 1], [0, 1, 1]]            => 64
  [[3, 0, 0], [0, 1, 2]]                       => 150
  [[1, 1, 0], [2, 0, 1], [0, 0, 1]]            => 312
  [[2, 0, 0], [1, 1, 1], [0, 0, 1]]            => 336
  [[2, 0, 0], [1, 1, 0], [0, 0, 2]]            => 242
  [[3, 0, 0], [0, 1, 1], [0, 0, 1]]            => 168
  [[3, 1, 1], [0, 0, 1]]                       => 432
  [[3, 0, 0], [0, 1, 0], [0, 0, 2]]            => 246
  [[0, 1, 0], [3, 0, 2]]                       => 108
  [[0, 1, 0], [2, 0, 1], [1, 0, 1]]            => 76
  [[1, 0, 0], [2, 0, 1], [0, 1, 1]]            => 122
  [[1, 0, 0], [2, 1, 1], [0, 0, 1]]            => 92
  [[2, 0, 0], [0, 1, 0], [1, 0, 2]]            => 312
  [[1, 0, 0], [2, 1, 2]]                       => 18
  [[2, 0, 0], [0, 1, 0], [1, 0, 1], [0, 0, 1]] => 132
  [[1, 0, 0], [1, 1, 1], [1, 0, 1]]            => 68
  ⋮                                            => ⋮

QuiverTools.all_destabilizing_subdimension_vectors — Function

Returns the subdimension vectors of $d$ with a strictly larger slope than $d$.

QuiverTools.all_generic_subdimension_vectors — Method

Returns the list of all generic subdimension vectors of $d$.

EXAMPLES:

julia> Q = mKronecker_quiver(3);

julia> QuiverTools.all_generic_subdimension_vectors(Q, [2, 3])
7-element Vector{AbstractVector{Int64}}:
 [0, 0]
 [0, 1]
 [0, 2]
 [1, 2]
 [0, 3]
 [1, 3]
 [2, 3]

julia> QuiverTools.all_generic_subdimension_vectors(Q, [3, 0])
4-element Vector{AbstractVector{Int64}}:
 [0, 0]
 [1, 0]
 [2, 0]
 [3, 0]

QuiverTools.all_slope_decreasing_sequences — Function

Returns the list of all sequences $(d^1,...,d^l)$ which sum to $d$ such that $\mu(d^1) > ... > \mu(d^l).$

EXAMPLES:

julia> Q = mKronecker_quiver(3); d = [2,3]; theta = [3,-2];

julia> QuiverTools.all_slope_decreasing_sequences(Q, d, theta)
8-element Vector{Vector{AbstractVector{Int64}}}:
 [[2, 3]]
 [[1, 1], [1, 2]]
 [[2, 2], [0, 1]]
 [[2, 1], [0, 2]]
 [[1, 0], [1, 3]]
 [[1, 0], [1, 2], [0, 1]]
 [[1, 0], [1, 1], [0, 2]]
 [[2, 0], [0, 3]]

QuiverTools.all_subdimension_vectors — Method

all_subdimension_vectors(d::AbstractVector{Int})

Compute all subdimension vectors of a given dimension vector d.

INPUT:

d::AbstractVector{Int}: The input dimension vector.

nonzero::Bool=false: wether to exclude the zero vector.

strict::Bool=false: wether to exclude the input vector d.

OUTPUT:

An array of all subdimension vectors of d, with or without the zero vector and d.

EXAMPLES:

julia> QuiverTools.all_subdimension_vectors([2, 3])
12-element Vector{AbstractVector{Int64}}:
 [0, 0]
 [1, 0]
 [2, 0]
 [0, 1]
 [1, 1]
 [2, 1]
 [0, 2]
 [1, 2]
 [2, 2]
 [0, 3]
 [1, 3]
 [2, 3]

julia> QuiverTools.all_subdimension_vectors([2, 3], nonzero=true)
11-element Vector{AbstractVector{Int64}}:
 [1, 0]
 [2, 0]
 [0, 1]
 [1, 1]
 [2, 1]
 [0, 2]
 [1, 2]
 [2, 2]
 [0, 3]
 [1, 3]
 [2, 3]

julia> QuiverTools.all_subdimension_vectors([2, 3], nonzero=true, strict=true)
10-element Vector{AbstractVector{Int64}}:
 [1, 0]
 [2, 0]
 [0, 1]
 [1, 1]
 [2, 1]
 [0, 2]
 [1, 2]
 [2, 2]
 [0, 3]
 [1, 3]

QuiverTools.all_weights_endomorphisms_universal_bundle — Function

Computes the weights of the endomorphisms of the universal bundles $U_i \otimes U_j$ on all the non-dense Harder-Narasimhan strata for each 1-PS relative to the HN type.

QuiverTools.all_weights_irreducible_component_canonical — Function

Computes the weights of the irreducible component of $\omega_R|_Z$ on all the non-dense Harder-Narasimhan strata for each 1-PS relative to the HN type. More explicitly, if $\omega_X = O(rH)$, this returns the weights of the pullback of $\mathcal{O}(H)$ on each stratum.

QuiverTools.all_weights_universal_bundle — Function

Computes the weights of the universal bundle $U_i(a)$ for the linearization $a$ on all the non-dense Harder-Narasimhan strata for each 1-PS corresponding to each HN type.

QuiverTools.arrows — Method

arrows(Q::Quiver)

Returns a list of all arrows of the quiver $Q$.

INPUT:

Q: a quiver

OUTPUT:

a list of all arrows of the quiver $Q$.

EXAMPLES:

julia> Q = mKronecker_quiver(3);

julia> arrows(Q)
3-element Vector{Vector{Int64}}:
 [1, 2]
 [1, 2]
 [1, 2]

QuiverTools.bipartite_quiver — Method

bipartite_quiver(m, n)

Constructs the bipartite quiver on m and n vertices.

EXAMPLES:

julia> bipartite_quiver(2, 3).adjacency
5×5 Matrix{Int64}:
 0  0  1  1  1
 0  0  1  1  1
 0  0  0  0  0
 0  0  0  0  0
 0  0  0  0  0

QuiverTools.canonical_decomposition — Method

Computes the canonical decomposition of the dimension vector $d$ for the given quiver $Q$.

If $\beta_1, \dots, \beta_{\ell}$ is a sequence of Schur roots such that, for all $i \neq j$, one has

\[\mathrm{ext}(\beta_i, \beta_j) = \mathrm{ext}(\beta_j, \beta_i) = 0,\]

then the general representation of dimension $\sum_i \beta_i$ is isomorphic to the direct sum of irreducible representations of dimension vectors $\beta_i$.

Such a decomposition is called the canonical decomposition.

EXAMPLES:

julia> Q = mKronecker_quiver(3);

julia> canonical_decomposition(Q, [6, 7]) == [[6, 7]]
true

julia> canonical_decomposition(Q, [1, 1]) == [[1, 1]]
true

julia> canonical_decomposition(Q, [6, 2]) == [[3, 1], [3, 1]]
true

julia> Q = mKronecker_quiver(2);

julia> canonical_decomposition(Q, [8, 8]) == [[1, 1] for i in 1:8]
true

QuiverTools.canonical_stability — Method

The canonical stability parameter for the couple $(Q, d)$ is given by $<d,-> - <-,d>$

INPUT:

Q: a quiver

d: a dimension vector

OUTPUT:

the canonical stability parameter for the couple $(Q, d)$

EXAMPLES:

julia> Q = mKronecker_quiver(3); d = [2,3];

julia> canonical_stability(Q, d) == [9, -6]
true

QuiverTools.codimension_HN_stratum — Method

Returns the codimension of the given HN stratum.

EXAMPLES:

julia> Q = mKronecker_quiver(3); d = [2,3]; theta = [3,-2];

julia> HN = all_HN_types(Q, d, theta)
8-element Vector{Vector{AbstractVector{Int64}}}:
 [[2, 3]]
 [[1, 1], [1, 2]]
 [[2, 2], [0, 1]]
 [[2, 1], [0, 2]]
 [[1, 0], [1, 3]]
 [[1, 0], [1, 2], [0, 1]]
 [[1, 0], [1, 1], [0, 2]]
 [[2, 0], [0, 3]]

julia> [codimension_HN_stratum(Q, stratum) for stratum in HN]
8-element Vector{Int64}:
  0
  3
  4
 10
  8
  9
 12
 18

QuiverTools.codimension_HN_stratum — Method

codimension_HN_stratum(M::QuiverModuli, hn_type::Vector{<AbstractVector{Int}})

Computes the codimension of the Harder-Narasimhan stratum corresponding to the given HN type.

INPUT:

M::QuiverModuli: a moduli space or stack of representations of a quiver.

hn_type::Vector{<:AbstractVector{Int}}: a HN type for M.

OUTPUT:

the codimension of the Harder-Narasimhan stratum corresponding to the given HN type.

EXAMPLES:

Codimensions for the 3-Kronecker quiver with dimension vector [2, 3]:

julia> Q = mKronecker_quiver(3); M = QuiverModuliSpace(Q, [2, 3]);

julia> codimension_HN_stratum(M, [[2, 3]])
0

julia> codimension_HN_stratum(M, [[1, 1], [1, 2]])
3

QuiverTools.codimension_unstable_locus — Method

codimension_unstable_locus(M::QuiverModuli)

Computes the codimension of the unstable locus in the parameter space.

INPUT:

M::QuiverModuli: a moduli space or stack of representations of a quiver.

OUTPUT:

the codimension of the unstable locus in the parameter space.

EXAMPLES:

Codimensions for the 3-Kronecker quiver with dimension vector [2, 3]:

julia> Q = mKronecker_quiver(3); M = QuiverModuliSpace(Q, [2, 3]);

julia> codimension_unstable_locus(M)
3

QuiverTools.coerce_to_quotient — Method

Takes a quotient ring R/I and a polynomial f in R and returns the image of f in R/I.

QuiverTools.cyclic_quiver — Method

cyclic_quiver(n)

Returns a cyclic quiver on n vertices.

EXAMPLES:

julia> cyclic_quiver(4)
cyclic quiver on 4 vertices, with adjacency matrix [0 1 0 0; 0 0 1 0; 0 0 0 1; 1 0 0 0]

QuiverTools.dimension — Method

dimension(M::QuiverModuliSpace)

Returns the dimension of the moduli space.

INPUT:

$M$: a moduli space of representations of a quiver.

OUTPUT:

the dimension of the moduli space.

EXAMPLES:

The dimension of the moduli space of the 3-Kronecker quiver:

julia> Q = mKronecker_quiver(3); M = QuiverModuliSpace(Q, [2, 3]);

julia> dimension(M)
6

QuiverTools.dimension — Method

dimension(M::QuiverModuliStack)

Returns the dimension of the moduli stack. This differs from the dimension of the moduli space by 1, as we do not quotient out the stabilizer $\mathbb{G}$.

INPUT:

$M$: a moduli stack of representations of a quiver.

OUTPUT:

the dimension of the moduli stack.

EXAMPLES:

The dimension of the moduli stack of the 3-Kronecker quiver

julia> Q = mKronecker_quiver(3); M = QuiverModuliStack(Q, [2, 3]);

julia> dimension(M)
5

QuiverTools.dimension_of_Luna_stratum — Method

dimension_of_Luna_stratum(M::QuiverModuli, tau)

Computes the dimension of the Luna stratum corresponding to the given Luna type in the moduli space.

INPUT:

M::QuiverModuli: a moduli space or stack of representations of a quiver.

tau::Dict{AbstractVector{Int}, Vector{Int}}: a Luna type for M.

OUTPUT:

the dimension of the Luna stratum corresponding to the given Luna type.

EXAMPLES:

julia> Q = mKronecker_quiver(2); M = QuiverModuliSpace(Q, [2, 2], [1, -1]);

julia> luna = all_Luna_types(M)
2-element Vector{Dict{AbstractVector, Vector{Int64}}}:
 Dict([1, 1] => [2])
 Dict([1, 1] => [1, 1])

julia> [dimension_of_Luna_stratum(M, tau) for tau in luna]
2-element Vector{Int64}:
 1
 2

QuiverTools.does_Teleman_inequality_hold — Method

does_Teleman_inequality_hold(M::QuiverModuli)

Checks if the Teleman quantization criterion of arXiv:2311.17003 holds.

In case the quiver is acyclic, this ensures that the moduli space is infinitesimally rigid.

QuiverTools.double_quiver — Method

The adjacency matrix of the double of a quiver is the sum of the adjacency matrix of the original quiver and its transpose.

EXAMPLES:

julia> Q = mKronecker_quiver();

julia> double_quiver(Q)
double of 2-Kronecker quiver, with adjacency matrix [0 2; 2 0]

QuiverTools.extended_gcd — Method

Computes the gcd and the Bezout coefficients of a list of integers.

INPUT:

$x$: a list of integers.

OUTPUT: A tuple containing:

the gcd of the integers,

a choice of Bezout coefficients.

EXAMPLES:

julia> QuiverTools.extended_gcd([2, 3, 4])
2-element Vector{Any}:
 1
  [-1, 1, 0]
  
julia> QuiverTools.extended_gcd([2, 3])
2-element Vector{Any}:
 1
  [-1, 1]

QuiverTools.generic_ext — Method

Computes the dimension of the $\mathrm{Ext}^1$ group between generic representations of dimension vectors $a$ and $b$.

According to Theorem 5.4, MR1162487, we have

\[ext(a,b)=max\{-\langle c,b\rangle~~|~~c~\text{is a generic subdimension vector of }a\}.\]

EXAMPLES:

julia> Q1 = mKronecker_quiver(3);

julia> generic_ext(Q1, [2, 3], [6, 7])
9

julia> generic_ext(Q1, [1, 1], [1, 0])
0

julia> Q2 = three_vertex_quiver(1, 6, 7);

julia> generic_ext(Q2, [5, 6, 7], [6, 7, 8])
483

QuiverTools.generic_hom — Method

Computes the dimension of the $\mathrm{Hom}$ group between generic representations of dimension vectors $a$ and $b$.

EXAMPLES:

julia> Q1 = mKronecker_quiver(3);

julia> generic_hom(Q1, [2, 3], [6, 7])
0

julia> generic_hom(Q1, [1, 1], [1, 0])
1

julia> Q2 = three_vertex_quiver(1, 6, 7);

julia> generic_hom(Q2, [5, 6, 7], [6, 7, 8])
0

QuiverTools.has_semistables — Function

Checks if there is a $\theta$-semistable representation of dimension vector $d$.

EXAMPLES:

julia> A2 = mKronecker_quiver(1); theta = [1,-1];

julia> has_semistables(A2, [1,1], theta)
true

julia> has_semistables(A2, [2,2], theta)
true

julia> has_semistables(A2, [1,2], theta)
false

julia> has_semistables(A2, [0,0], theta)
true

The 3-Kronecker quiver:

julia> K3 = mKronecker_quiver(3); theta = [3,-2];

julia> has_semistables(K3, [2,3], theta)
true

julia> has_semistables(K3, [1,4], theta)
false

QuiverTools.has_stables — Function

Checks if Q has a $theta$-stable representation of dimension vector $d$.

EXAMPLES:

julia> Q = mKronecker_quiver(3); d = [2, 3]; theta = [3, -2];

julia> has_stables(Q, d, theta)
true

julia> Q = mKronecker_quiver(2); d = [2,2]; theta = [1,-1];

julia> has_stables(Q, d, theta)
false

julia> has_semistables(Q, d, theta)
true

The zero dimension vector has no stables:

julia> Q = mKronecker_quiver(3); d = [0,0];

julia> has_stables(Q, d)
false

QuiverTools.identity_matrix — Method

Returns the identity matrix of size $n$.

QuiverTools.in_fundamental_domain — Method

Checks if the dimension vector $d$ is in the fundamental domain of the quiver $Q$.

The fundamental domain is the cone of dimension vectors in $\mathbb{Z}^{Q_0}$ such that the symmetric Tits form is negative on all the simple roots, i.e., for all vertices i,

\[(s_i, d) := \langle d, s_i\rangle + \langle s_i, d\rangle \leq 0,\]

where $s_i$ is the dimension vector with all entries set to $0$ and the i-th set to $1$.

EXAMPLES:

julia> Q = mKronecker_quiver(3);

julia> in_fundamental_domain(Q, [2, 3])
true

julia> in_fundamental_domain(Q, [1, 1])
true

julia> in_fundamental_domain(Q, [2, 2])
true

julia> in_fundamental_domain(Q, [1, 2])
false

QuiverTools.indegree — Method

Returns the number of incoming arrows to the vertex $j$.

EXAMPLES:

julia> Q = mKronecker_quiver(4);

julia> indegree(Q, 1)
0

julia> indegree(Q, 2)
4

QuiverTools.index — Method

index(M::QuiverModuliSpace)

Returns the index of the moduli space $M$.

The index of a variety $X$ is the largest which divides the canonical divisor $K_X$ in $Pic(X)$.

This implementation currently only works for the canonical stability.

INPUT:

$M$: a moduli space of representations of a quiver.

OUTPUT:

the index of the moduli space.

EXAMPLES:

The 3-Kronecker quiver has index 3:

julia> Q = mKronecker_quiver(3);

julia> M = QuiverModuliSpace(Q, [2, 3]);

julia> index(M)
3

The subspace quiver moduli have index 1:

julia> Q = subspace_quiver(5);

julia> M = QuiverModuliSpace(Q, [1, 1, 1, 1, 1, 2]);

julia> index(M)
1

QuiverTools.integral — Function

integral(M, f, chi)

Computes the integral of f according to the Hirzebruch-Riemann-Roch theorem.

In other words, it computes the Euler characteristic of the vector bundle whose Chern character is f.

INPUT:

$M$: a moduli space of representations of a quiver.

f: the Chern character in CH(M) to integrate.

$chi$: a choice of linearization to construct the universal bundles.

OUTPUT: the integral of f.

EXAMPLES:

The integral of $\mathcal{O}(i)$ on the projective line for some is.

julia> Q = mKronecker_quiver(2); M = QuiverModuliSpace(Q, [1, 1]);

julia> L = Chern_character_line_bundle(M, [1, -1]);

julia> [integral(M, L^i) for i in 0:5]
6-element Vector{Int64}:
 1
 2
 3
 4
 5
 6

Hilbert series for the 3-Kronecker quiver as in our favourite 6-fold:

julia> Q = mKronecker_quiver(3); M = QuiverModuliSpace(Q, [2, 3]);

julia> L = Chern_character_line_bundle(M, [3, -2]);

julia> [integral(M, L^i) for i in 0:5]
6-element Vector{Int64}:
    1
   20
  148
  664
 2206
 5999

QuiverTools.is_HN_type — Function

is_hn_type(Q, d, dstar, theta, denom)

Checks if the given ordered list of subdimension vectors $dstar$ is an HN type for the datum $(Q, d)$ and the slope stability given by $(theta, denom)$.

EXAMPLES:

julia> Q = mKronecker_quiver(3);

julia> d = [2, 3]; dstar = [d];

julia> is_HN_type(Q, d, dstar)
true

QuiverTools.is_HN_type — Method

is_HN_type(M::QuiverModuli, hn_type::AbstractVector{<:AbstractVector{Int}})

Checks if the given sequence of dimension vectors is a valid HN type for the moduli space.

INPUT:

M::QuiverModuli: a moduli space or stack of representations of a quiver.

hn_type::AbstractVector{<:AbstractVector{Int}}: a sequence of dimension vectors.

OUTPUT:

whether the given sequence is a valid Harder-Narasimhan type for M.

EXAMPLES:

Some HN types for the 3-Kronecker quiver with dimension vector [2, 3]:

julia> Q = mKronecker_quiver(3); M = QuiverModuliSpace(Q, [2, 3]);

julia> is_HN_type(M, [[2, 3]])
true

julia> is_HN_type(M, [[1, 1], [1, 2]])
true

julia> is_HN_type(M, [[1, 2], [1, 1]])
false

QuiverTools.is_Luna_type — Method

is_Luna_type(M::QuiverModuli, tau)

Checks if the given tau is a valid Luna type for M.

INPUT:

M::QuiverModuli: a moduli space or stack of representations of a quiver.

tau::Dict{AbstractVector{Int}, Vector{Int}}: a Luna type for M.

OUTPUT:

whether the given tau is a valid Luna type for M.

EXAMPLES:

Nontrivial Luna types for the 3-Kronecker quiver:

julia> Q = mKronecker_quiver(3); M = QuiverModuliSpace(Q, [3, 3]);

julia> l = Dict([1, 1] => [1], [2, 2] => [1]);

julia> is_Luna_type(M, l)
true

The zero dimensional case:

julia> Q = mKronecker_quiver(3); X = QuiverModuliSpace(Q, [0, 0]);

julia> is_Luna_type(X, Dict([0, 0] => [1]))
true

QuiverTools.is_Schur_root — Method

Checks if $d$ is a Schur root for $Q$.

By Lemma 4.2, arXiv:0802.2147, this is equivalent to the existence of a stable representation of dimension vector $d$ for the canonical stability parameter.

EXAMPLES:

julia> Q = mKronecker_quiver(3); d = [2,3];

julia> is_Schur_root(Q, d)
true

QuiverTools.is_acyclic — Method

Checks wether the quiver is acyclic, i.e. has no oriented cycles.

julia> Q = mKronecker_quiver(4);

julia> is_acyclic(Q)
true

QuiverTools.is_amply_stable — Function

Checks wether the dimension vector $d$ is amply stable with respect to the slope function theta/denominator.

This means that the codimension of the unstable locus in the parameter space is at least $2$.

EXAMPLES:

julia> Q = mKronecker_quiver(3); d = [2, 3];

julia> is_amply_stable(Q, d, [3, -2])
true

julia> is_amply_stable(Q, d, [-3, 2])
false

julia> is_amply_stable(Q, [3, 0], [0, -3])
true

QuiverTools.is_amply_stable — Method

is_amply_stable(M::QuiverModuli)

Checks wether the dimension vector $d$ is amply stable with respect to the slope function theta/denominator.

This means that the codimension of the unstable locus in the parameter space is at least $2$.

INPUT:

M::QuiverModuli: a moduli space or stack of representations of a quiver.

OUTPUT:

whether the codimension of the unstable locus is at least 2.

EXAMPLES:

julia> Q = mKronecker_quiver(3); M = QuiverModuliSpace(Q, [2, 3]);

julia> is_amply_stable(M)
true

QuiverTools.is_connected — Method

Checks wether the underlying graph of the quiver is connected.

EXAMPLES:

julia> Q = Quiver([0 1 0; 0 0 1; 1 0 0]);

julia> is_connected(Q)
true

julia> Q = Quiver([0 1 0; 1 0 0; 0 0 2]);

julia> is_connected(Q)
false

julia> # The 4-Kronecker quiver:

julia> Q = mKronecker_quiver(4);

julia> is_connected(Q)
true

julia> # The 4-loop quiver:

julia> Q = loop_quiver(4);

julia> is_connected(Q)
true

julia> # The 4-subspace quiver:

julia> Q = subspace_quiver(4);

julia> is_connected(Q)
true

QuiverTools.is_coprime — Method

Checks wether the given dimension vector $d$ is $\theta$-coprime for the stability parameter $\theta$.

EXAMPLES:

julia> d = [2, 3]; theta = [3, -2];

julia> is_coprime(d, theta)
true

QuiverTools.is_coprime — Method

Checks if the gcd of all the entries of d is $1$.

QuiverTools.is_coprime — Method

is_coprime(M::QuiverModuli)

Checks if the stability parameter is coprime with the dimension vector, i.e., if for all subdimension vectors $e$ of $d$, $\theta\cdot e \neq 0$.

INPUT:

M::QuiverModuli: a moduli space or stack of representations of a quiver.

OUTPUT:

whether the dimension vector M.d is theta-coprime for M.theta.

EXAMPLES:

julia> Q = mKronecker_quiver(3);

julia> M = QuiverModuliSpace(Q, [2, 3]);

julia> is_coprime(M)
true

QuiverTools.is_generic_subdimension_vector — Method

Checks if $e$ is a generic subdimension vector of $d$.

A dimension vector $e$ is called a generic subdimension vector of $d$ if a generic representation of dimension vector $d$ possesses a subrepresentation of dimension vector $e$.

By Theorem 5.3, arXiv:0802.2147, $e$ is a generic subdimension vector of $d$ if and only if

\[<e',d-e> \geq 0\]

for all generic subdimension vectors $e'$ of $e$.

QuiverTools.is_nonempty — Method

is_nonempty(M::QuiverModuli)

Checks if the quiver moduli is nonempty.

INPUT:

M::QuiverModuli: a moduli space or stack of representations of a quiver.

OUTPUT:

whether the moduli space is nonempty.

EXAMPLES:

julia> Q = mKronecker_quiver(3);

julia> M = QuiverModuliSpace(Q, [2, 3]);

julia> is_nonempty(M)
true

julia> M = QuiverModuliSpace(Q, [2, 3], [-3, 2]);

julia> is_nonempty(M)
false

QuiverTools.is_projective — Method

is_projective(M::QuiverModuli)

Checks if the moduli space is projective.

INPUT:

$M$: a moduli space of representations of a quiver.

OUTPUT:

whether the moduli space is projective.

EXAMPLES:

The moduli space of the 3-Kronecker quiver is projective:

julia> Q = mKronecker_quiver(3); M = QuiverModuliSpace(Q, [2, 3]);

julia> is_projective(M)
true

QuiverTools.is_sink — Method

Checks if the vertex $j$ is a sink, i.e., a vertex with no outgoing arrows.

EXAMPLES:

julia> Q = mKronecker_quiver(4);

julia> is_sink(Q, 1)
false

julia> is_sink(Q, 2)
true

QuiverTools.is_smooth — Method

is_smooth(M::QuiverModuliSpace)

Checks if the moduli space is smooth.

INPUT:

$M$: a moduli space of representations of a quiver.

OUTPUT:

whether the moduli space is smooth.

EXAMPLES:

Setups with d theta-coprime are smooth:

julia> Q = mKronecker_quiver(3); M = QuiverModuliSpace(Q, [2, 3]);

julia> is_smooth(M)
true

QuiverTools.is_smooth — Method

is_smooth(M::QuiverModuliStack)

Checks if the moduli stack is smooth.

This is always trus, as the quotient stack of a smooth variety is smooth.

INPUT:

$M$: a moduli stack of representations of a quiver.

OUTPUT:

true

EXAMPLES:

julia> Q = mKronecker_quiver(3); M = QuiverModuliStack(Q, [2, 3]);

julia> is_smooth(M)
true

QuiverTools.is_source — Method

Checks if the vertex $i$ is a source, i.e., a vertex with no incoming arrows.

EXAMPLES:

julia> Q = mKronecker_quiver(4);

julia> is_source(Q, 1)
true

julia> is_source(Q, 2)
false

QuiverTools.is_subdimension_vector — Method

is_subdimension_vector(e::AbstractVector{Int}, d::AbstractVector{Int})

Check if vector e is a subdimension of vector d.

INPUT:

e: An abstract vector of integers.

d: An abstract vector of integers.

OUTPUT: whether e is a subdimension of d.

EXAMPLE:

julia> QuiverTools.is_subdimension_vector([1, 1], [2, 3])
true

julia> QuiverTools.is_subdimension_vector([1, 1], [1, 1])
true

julia> QuiverTools.is_subdimension_vector([1, 2], [1, 1])
false

QuiverTools.local_quiver_setting — Method

local_quiver_setting(M::QuiverModuli, tau)

Returns the local quiver and dimension vector for the given Luna type.

INPUT:

M::QuiverModuli: a moduli space or stack of representations of a quiver.

tau::Dict{AbstractVector{Int}, Vector{Int}}: a Luna type for M.

OUTPUT:

a dictionary with the local quiver Q and dimension vector d for the given Luna type.

QuiverTools.loop_quiver — Method

loop_quiver(m::Int)

Constructs a loop quiver with m vertices.

INPUT:

m: The number of vertices in the loop quiver.

OUTPUT: A loop quiver with m vertices.

EXAMPLES:

julia> loop_quiver(4)
4-loop quiver, with adjacency matrix [4;;]

QuiverTools.mKronecker_quiver — Function

mKronecker_quiver(m::Int)

Constructs a Kronecker quiver with m vertices.

INPUT:

m: (Default = 2) The number of arrows in the Kronecker quiver.

OUTPUT: A Kronecker quiver with m vertices.

EXAMPLES:

julia> mKronecker_quiver(3)
3-Kronecker quiver, with adjacency matrix [0 3; 0 0]

QuiverTools.motive — Function

motive(Q, d, theta, denom)

Returns the motive of the moduli stack of $\theta$-semistable representations.

INPUT:

$Q$: a quiver.

$d$: a dimension vector.

$theta$: a stability parameter. Default is the canonical stability.

$denom$: a function. Default is the sum.

OUTPUT:

The motive as an element in the function field \mathbb{Q}(L).

EXAMPLES:

julia> Q = mKronecker_quiver(3);

julia> motive(Q, [2, 3])
(-L^6 - L^5 - 3*L^4 - 3*L^3 - 3*L^2 - L - 1)//(L - 1)

QuiverTools.narrows — Method

Returns the number of arrows of the quiver.

julia> Q = mKronecker_quiver(4);

julia> narrows(Q) == 4
true

QuiverTools.nvertices — Method

Returns the number of vertices of the quiver.

julia> Q = mKronecker_quiver(4);

julia> nvertices(Q) == 2
true

QuiverTools.opposite_quiver — Method

" Returns a Quiver with the same vertices and an arrow $j \to i$ for every arrow $i \to j$ in the original quiver.

INPUT:

Q::Quiver: The quiver to be reversed.

OUTPUT: A quiver with the same vertices and reversed arrows.

EXAMPLES:

julia> Q = mKronecker_quiver()
2-Kronecker quiver, with adjacency matrix [0 2; 0 0]

julia> opposite_quiver(Q)
opposite of 2-Kronecker quiver, with adjacency matrix [0 0; 2 0]

QuiverTools.outdegree — Method

Returns the number of outgoing arrows from the vertex $i$.

EXAMPLES:

julia> Q = mKronecker_quiver(4);

julia> outdegree(Q, 1)
4

julia> outdegree(Q, 2)
0

QuiverTools.point_class — Function

point_class(M::QuiverModuliSpace)

Returns the point class of the moduli space $M$.

INPUT:

$M$: a moduli space of representations of a quiver.

$chi$: a choice of linearization to construct the universal bundles.

OUTPUT:

the point class of the moduli space, as a polynomial in its Chow ring.

EXAMPLES:

A projective 7-fold:

julia> Q = mKronecker_quiver(8);

julia> M = QuiverModuliSpace(Q, [1, 1]);

julia> point_class(M, [1, 0])
x21^7

Our favourite 6-fold:

julia> Q = mKronecker_quiver(3);

julia> M = QuiverModuliSpace(Q, [2, 3]);

julia> point_class(M)
x23^2

QuiverTools.pullback_from_quotient — Method

Takes a ring R and a polynomial in R/I and returns the canonical preimage of f in R.

QuiverTools.semisimple_moduli_space — Method

semisimple_moduli_space(M::QuiverModuli)

Returns the moduli space with the zero stability parameter.

Any quiver moduli space is (quasi)projective-over-affine; this is the affine base.

INPUT:

$M$: a moduli space of representations of a quiver.

OUTPUT:

the moduli space with the zero stability parameter.

EXAMPLES:

julia> Q = mKronecker_quiver(3); M = QuiverModuliSpace(Q, [2, 3]);

julia> dimension(semisimple_moduli_space(M))
0

QuiverTools.semistable_equals_stable — Method

semistable_equals_stable(M::QuiverModuli)

Checks if stability and semistability are equivalent on the given moduli space. In other words, checks if there are no properly semistable points in the representation space.

INPUT:

M::QuiverModuli: a moduli space or stack of representations of a quiver.

OUTPUT:

whether every semistable representation is stable.

EXAMPLES:

If the dimension vector is coprime with the stability parameter, then semistability and stability are equivalent:

julia> Q = mKronecker_quiver(3); M = QuiverModuliSpace(Q, [2, 3]);

julia> semistable_equals_stable(M)
true

However, this is not necessarily the case:

julia> Q = mKronecker_quiver(3); M = QuiverModuliSpace(Q, [3, 3]);

julia> semistable_equals_stable(M)
false

QuiverTools.slope — Function

Returns the slope of the dimension vector $d$ with respect to the stability parameter $\theta$ and a choice of a denominator function.

EXAMPLE:

julia> slope([2,3], [3,-2])
0//1

QuiverTools.solve — Method

solve(A, b)

Solve $A\cdot x = b$ for $A$ upper triangular via back substitution.

This is an internal method only used in the implementation of the Hodge polynomial and to compute motives.

INPUT:

A::AbstractMatrix: an upper triangular matrix.

b::AbstractVector: a vector.

OUTPUT:

the solution x to the equation.

EXAMPLES:

julia> A = [1 2 3; 0 4 5; 0 0 6];

julia> b = [1, 2, 3];

julia> QuiverTools.solve(A, b)
3-element Vector{Any}:
 -0.25
 -0.125
  0.5

QuiverTools.subspace_quiver — Method

subspace_quiver(m::Int)

Constructs a subspace quiver with m vertices.

INPUT:

m: The number of subspace-vertices.

OUTPUT: A subspace quiver with m subspaces.

EXAMPLES:

julia> subspace_quiver(3)
3-subspace quiver, with adjacency matrix [0 0 0 1; 0 0 0 1; 0 0 0 1; 0 0 0 0]

QuiverTools.symmetric_polynomial — Method

symmetric_polynomial(vars, degree)

Returns the symmetric polynomial of degree $degree$ in the variables $vars$.

INPUT:

$vars$: a list of variables.

$degree$: the degree of the wanted symmetric polynomial.

OUTPUT:

The symmetric polynomial of degree $degree$ in the variables $vars$.

EXAMPLES:

julia> using Singular;

julia> R, vars = polynomial_ring(Singular.QQ, ["x", "y", "z"]);

julia> QuiverTools.symmetric_polynomial(vars, 2)
x*y + x*z + y*z

QuiverTools.thin_dimension_vector — Method

thin_dimension_vector(Q::Quiver)

Compute the thin dimension vector for a given quiver Q.

INPUT:

Q::Quiver: The input quiver.

OUTPUT:

A vector of ones of length n.

EXAMPLE:

There is not much to it:

julia> Q = mKronecker_quiver(3);

julia> QuiverTools.thin_dimension_vector(Q) == [1, 1]
true

QuiverTools.three_vertex_quiver — Method

three_vertex_quiver(m12::Int, m13::Int, m23::Int)

Constructs a three-vertex quiver with the given edge weights.

INPUT:

m12: The number of arrows from vertex 1 to vertex 2.

m13: The number of arrows from vertex 1 to vertex 3.

m23: The number of arrows from vertex 2 to vertex 3.

OUTPUT: A three-vertex quiver with the specified arrows.

EXAMPLES:

julia> three_vertex_quiver(1, 2, 3)
Acyclic 3-vertex quiver, with adjacency matrix [0 1 2; 0 0 3; 0 0 0]

QuiverTools.total_Chern_class_universal — Function

total_Chern_class_universal(M::QuiverModuliSpace, i, chi)

Returns the total Chern class of the universal bundle $U_i(\chi)$.

INPUT:

$M$: a moduli space of representations of a quiver.

$i$: the universal bundle we want the Chern class of.

$chi$: a choice of linearization to construct $U_i(\chi)$.

OUTPUT:

the total Chern class of the universal bundle $U_i(\chi)$.

EXAMPLES:

The universal Chern classes on both vertices of our favourite 3-Kronecker quiver:

julia> Q  = mKronecker_quiver(3); M = QuiverModuliSpace(Q, [2, 3]);

julia> total_Chern_class_universal(M, 1)
x11 + x12 + 1

julia> total_Chern_class_universal(M, 2)
x21 + x22 + x23 + 1

QuiverTools.underlying_graph — Method

Returns the (necessarily symmetric) adjacency matrix of the underlying graph of the quiver.

julia> Q = mKronecker_quiver(4);

julia> QuiverTools.underlying_graph(Q) == [0 4; 4 0]
true

QuiverTools.unit_vector — Method

unit_vector(n::Int, i::Int)

Return a vector of length n with a 1 at index i and 0 elsewhere.

INPUT:

n::Int: The length of the unit vector.

i::Int: The index at which to place the 1 in the unit vector.

OUTPUT: A unit vector of length n with a 1 at index i and 0 elsewhere.

EXAMPLES:

julia> QuiverTools.unit_vector(3, 2) == [0, 1, 0]
true

QuiverTools.unit_vector — Method

unit_vector(Q::Quiver, i::Int)

Return a dimension vector for the quiver Q with a 1 at index i and 0 elsewhere.

INPUT:

Q::Quiver: The input quiver.

i::Int: The index at which to place the 1 in the unit vector.

OUTPUT:

A dimension vector for the quiver Q with a 1 at index i and 0 elsewhere.

EXAMPLES:

julia> Q = mKronecker_quiver(3);

julia> QuiverTools.unit_vector(Q, 2) == [0, 1]
true

QuiverTools.weight_irreducible_component_canonical_on_stratum — Function

Computes the weight of the irreducible component of $\omega_R|_Z$ on a Harder-Narasimhan stratum for the 1-PS corresponding to each HN type. More explicitly, if $\omega_X = \mathcal{O}(rH)$, this returns the weight of the pullback of O(H) on the given stratum.

QuiverTools.weights_endomorphism_universal_bundle_on_stratum — Function

Computes the weights of the endomorphism of the universal bundle $U_i \otimes U_j$ on the given Harder-Narasimhan stratum for the 1-PS relative to the HN type.

QuiverTools.weights_universal_bundle_onstratum — Function

Returns the weights of a universal bundle $U_i(a)$ for the linearization $a$ for the 1-PS corresponding to the given HN type.

QuiverTools.zero_vector — Method

zero_vector(n::Int)

Create a zero vector of length n.

INPUT:

n::Int: The length of the zero vector.

OUTPUT:

A zero vector of length n.

EXAMPLE:

There is not much to it:

julia> QuiverTools.zero_vector(3) == [0, 0, 0]
true